BaseTodo/requirements.txt
anyio==3.7.1
asgiref==3.10.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
gunicorn==20.1.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
motor==3.7.1
passlib==1.7.4
pycparser==2.23
pydantic==1.10.24
PyJWT==2.3.0
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-decouple==3.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.16.0
style==1.1.0
typing_extensions==4.15.0
update==0.0.1
uvicorn==0.38.0
